requirements.txt
# Project requirements for memory map visualization and graph computation
matplotlib
umap-learn
numpy
pytest
